cleanup: add exit code as parameter

+++ lazybar-core/src/cleanup.rs
@@ -1,6 +1,8 @@
 use std::fs::remove_file;
 
-pub fn exit(bar: &str) {
-    let _ = remove_file(format!("/tmp/lazybar-ipc/{bar}"));
-    std::process::exit(0);
+pub fn exit(bar: Option<&str>, exit_code: i32) -> ! {
+    if let Some(bar) = bar {
+        let _ = remove_file(format!("/tmp/lazybar-ipc/{bar}"));
+    }
+    std::process::exit(exit_code);
 }
